requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
rdkit-pypi==2022.9.5
numpy==1.24.3
pydantic==2.5.0
redis==4.6.0
celery[redis]==5.3.4
flower==2.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
flake8==6.1.0
black==23.11.0
mypy==1.7.1
pre-commit==3.5.0